refactor: spinnerコンポーネントを分離
